requirements.txt
amqp==5.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.4.0
certifi==2023.7.22
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==5.1.5
django-cors-headers==4.6.0
django-csp==4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.1.0
requests==2.31.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.1.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zarinpal==1.0.0
zope.interface==7.2
